feat(protocol-contracts): add deployment and verification scripts for governance
